Bundle Selleckt legacy shim into main browserified package.
The selleckt shim code is pretty small, and so it makes sense to keep
it bundled within the main package. By doing this, we remove the onus
from the package consumer to include the separate legacy shim.

This has the added advantage of cleaning up code for AMD-based consumers
who may have had to dynamically construct requires as so:

```
var libs = window.MutationObserver && window.MutationObserver._period ?
            ['selleckt-legacy'] :
            ['selleckt'];

define(libs, function(selleckt) {
    return selleckt;
});
```

Now we just check inside the main bundle for the existence of the
window.MutationObserver shim, and if it does exist we make the
appropriate changes to Selleckt